MenuService and MenuController
MenuService and MenuController methods were implemented

+++ StoreHouse.Api/Model/DTO/ProductListDTO/MenuProductListResponse.cs
@@ -5,5 +5,5 @@ public class MenuProductListResponse
     public int Id { get; set; }
     public string Name { get; set; }
     public double Weight { get; set; }
-    public double PrimeCost { get; set; }
+    public decimal PrimeCost { get; set; }
 }
